Add files via upload
I'd actually forgotten about this version of the game and it'd sat in a private discord message with a friend for a couple months. It's here and it's actually pretty cool with an entire loot and inventory system, just came back to the project and am probably going to re-make it. I need to figure out how I can efficiently allow the user to return to the adventure menu when they're finished with looking through their inventory.